% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_missing_interval_fn.R
\name{find_missing_index}
\alias{find_missing_index}
\title{Identify and Document Missing Data Intervals in a Data Frame.}
\usage{
find_missing_index(input_vec)
}
\arguments{
\item{input_vec}{A numeric data vector.}
}
\value{
A data frame with three columns:
\itemize{
\item \code{start_missing_index}: Starting positions of missing data intervals.
\item \code{end_missing_index}: Ending positions of missing data intervals.
\item \code{missing_period}: Lengths of the missing data intervals.
}
}
\description{
Identify and Document Missing Data Intervals in a Data Frame.
}
\examples{
find_missing_index(c(1, 2, 3, NA, NA, 6, 7))

}
