% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_interpolation_fn.R
\name{linear_interpolation}
\alias{linear_interpolation}
\title{Apply linear interpolation to fill missingness of a specific interval in data.}
\usage{
linear_interpolation(data_vec, start_missing_indices, end_missing_indices)
}
\arguments{
\item{data_vec}{A numeric data vector.}

\item{start_missing_indices}{The position where the missingness interval begins.}

\item{end_missing_indices}{The position where the missingness interval ends.}
}
\value{
A numeric vector with the specified missing interval filled.
}
\description{
Apply linear interpolation to fill missingness of a specific interval in data.
}
\examples{
linear_interpolation(c(1, 2, 3, NA, NA, 6, 7), 4, 5)
}
