% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_linear_interpolation_fn.R
\name{linear_interpolation_df}
\alias{linear_interpolation_df}
\title{Apply linear interpolation to fill missing values in specified columns of a dataframe.}
\usage{
linear_interpolation_df(df, col_names, cut_off)
}
\arguments{
\item{df}{A data frame containing numeric data.}

\item{col_names}{A character vector of column names to process.}

\item{cut_off}{The maximum length of consecutive missing values that can be filled with linear interpolation.}
}
\value{
A data frame with missing values in the specified columns filled using linear interpolation.
}
\description{
Apply linear interpolation to fill missing values in specified columns of a dataframe.
}
